xchanger_requirements.txt
logging_exceptions
python-dotenv
pika
requests
pyyaml
munch
tenacity
bugsnag
